moved gammaln
